ps: firmware already handle powersave mode during scan
When user try to launch scan while connected, it is necessary to notify
the AP that we cannot receive data (using power save mode).

Firmware already handle this automatically so the code in driver is
redundant and can be dropped.

By edge effect, hack of scan status in wfx_set_pm() is now useless.