ps: drop unnecessary wvif->powersave_mode
Power save status is already available in bss_conf. So there is no
reason to keep information duplicated in wvif->powersave_mode.

In add, type of wvif->powersave_mode is low level struct made to
communicate with device. We would like to limit usage of this kind of
struct in upper layers of the driver.