ps: simplify hif_set_uapsd_info() usage
It is useless to keep uapsd_info in struct wfx_vif. This structure can
be rebuilt just before to be sent.

In add, the struct hif_mib_set_uapsd_information comes from hardware
API. It is not intended to be manipulated in upper layers of the driver.
So, this patch relocate the handling of this struct to
hif_set_uapsd_info() (the low level function).