tx: fix the cache of rate policies on interface reset
Device and driver maintain a cache of rate policies (aka.
tx_retry_policy in hardware API).

When hif_reset() is sent to hardware, device reset its cache of rate
policies. In order to keep driver in sync, it is necessary to do the
same on driver.

Note, when driver try to use a rate policy that has not been defined on
device, data is sent to 1Mbps. So, this patch should fix abnormal
throughput observed sometime after a reset of the interface.