tx: fix rate control handling
A tx_retry_policy (the equivalent of ieee80211_tx_rate list in hardware
API) is not able to include a rate multiple time. So, the driver merge
the identical rates from the policy provided by minstrel (and it try to
do the best choice it can in the associated flags).

Until now, when rate are merged, count field is set to
max(count1, count2). But, it means that total number of retry could be
far less than initial number of retry. So change value of count field to
count1 + count2.