sta: change the way the station associate to an AP
Chipset need two steps to associate with an AP:
   1. it start receiving beacon from the AP (done with wfx_do_join())
   2. it sent the association request (done with wfx_join_finalize())

The join request (see hif_join()) contains basic rates, beacon interval
and bssid to connect, so we trig on these events for the first step.

The second step is obviously associated to the event BSS_CHANGED_ASSOC.

Note that conf_mutex is now easier to manage. It is held by
wfx_bss_info_changed() and inner functions do need to lock it.