hif: fix bytes order
wfx_sl_check_pubkey() aims to check the cryptographic key contained in
the chip. The code for this job exists[1], but is not yet integrated in
mainstream since it pull external dependencies (see related item in TODO
file included beside wfx driver).

Anyway, the commit 36f7e3acaac6 ("staging: wfx: ensure that received hif
messages are never modified") reversed the order of the bytes of pubkey.
The goal was to simplify the job of wfx_sl_check_pubkey().

However, this does not work. wfx_sl_check_pubkey() need the original
value of pubkey. So revert that change.

[1]: https://github.com/SiliconLabs/wfx-linux-driver/blob/master/secure_link.c

Fixes: 36f7e3acaac6 ("staging: wfx: ensure that received hif messages are never modified")