sta: fix CAB sent at the wrong time
wfx_suspend_resume_mc() is called when the device is about to sent a
DTIM. This is the right moment to enqueue Content After DTIM Beacon
(CAB).

However, wfx_suspend_resume_mc() is also called when the DTIM period
ends. Until now, this event did also trig CAB.

Note this issue did not have too much impact since when a CAB is sent
outside of DTIM window, an error is reported by the firmware and
mac80211 retries to send the data.