tx: allow to disable QoS in output frames
Slow platforms may have some problems to pass WiFi Alliance (WFA)
certification because they are not fast enough to process all the
traffic (or the bus is not fast enough to transfer all data). They
generally encounter difficulties in QoS tests.

One way to workaround this problem is to declare itself as
Application-Specific Devices (ASD) that does not support QoS. In this
case, the QoS must be completely disabled to comply with the
certification.

Thus, this patch allows to disable the QoS on driver probing with:

    modprobe wfx disable_qos=1

It also possible to enable/disable QoS dynamically at runtime with:

    echo 1 > /sys/module/wfx/parameters/disable_qos

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>