tx: fix possible panic with re-queued frames (IPWHIFER-665)
When the firmware reject a frames (because station become asleep or
disconnected), the frame is re-queued in mac80211. However, the
re-queued frame was 8 bytes longer than the original one (the size of
the ICV for the encryption). So, when mac80211 try to send this frame
again, it is a little bigger than expected.
If the frame is re-queued secveral time it end with a skb_over_panic
because the skb buffer is not large enough.

This problem has been introduced by commit 049fde130419 ("staging: wfx:
drop useless field from struct wfx_tx_priv"). It only happens when
device acts as an AP and encryption is enabled.

Fixes: 049fde130419 ("staging: wfx: drop useless field from struct wfx_tx_priv")