debug: improve debug message on PS error
The WF200 tries hard to minimize power consumption. Unfortunately, when
the remote AP does not play fair, the WF200 may have to disable its
power saving optimizations. The driver report the warning to the user.

The current code show a scary message that trigs many support requests.
This patch provide better descriptions of these event and highlight it
only impacts the power consumption.