probe: improve user experience
The firmware of the WF200 is splited in two parts:
   - The chip vendor provides firmware
   - The card vendor provides "PDS" files (which in fact ta
     configuration file containing antenna configuration[1]).

Until now, the 'compatible' string was 'silabs,wf200' and the user had
to change the PDS file depending on the antenna he has. Now, the user
fix the 'compatible' to the name of the card (chip + antenna). We
already provide configuration for developments boards provided by Silabs
but it is easy to add custom cards[2].

[1]: https://github.com/SiliconLabs/wfx-firmware/tree/FW3.12.3/PDS
[2]: https://github.com/SiliconLabs/wfx-pds

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>