sta: allow to send frames during ROC
Until now, all the traffic was blocked during Remain On Channel (ROC).
This patch allow to handle IEEE80211_TX_CTL_TX_OFFCHAN frames.

These frames need to be sent on the virtual interface #2. Until now,
this interface was only used by the device for internal purpose. But
since API 3.9, it can be used to send data during scan operation (we
hijack the scan process to implement ROC).

Thus, we need to change a bit the way we match the frames with the
interface.

Fortunately, the frames received during the scan are marked with the
correct interface number. So there is no change to do on this part.