scan: scan_lock is global to the device
Currently, one scan_lock is associated to each vif. However, concurrent
scan on vifs is explicitly prohitibed by the device. Currently,
scan_lock is associated with conf_mutex that ensure that concurrent scan
on vifs cannot happen.

In the only case where conf_mutex is not associated to scan_lock, the
scan_lock is tested on all interfaces.

So, this patch relocate scan_lock to the device and simplify the code.